Test moved to separate file.
